[scanner] fix: resolve TS build errors from module split
- dashboardState.actions.ts: relax MoveToDashboardDeps.moveCardToDashboard
  return type to Promise<unknown> and type ConfirmDeployDeps.publishCardEvent
  as CardEvent instead of a loose object type (TS2345/TS2322, inherited
  from main).
- useClusterPageState.ts: type the injected t dependency as TFunction
  instead of a loose (key, opts) => string signature, fixing the
  Clusters.tsx TS2322 assignment error.
- Clusters.tsx: drop the unused ClusterLayoutMode import (TS6133).
- ResourceTreeView.tsx: align local Pod/Deployment/Service/Job/HPA/
  ConfigMap/Secret/ServiceAccount/PVC field types with the optional
  fields on the real hooks/mcp types (TS2322), and stop destructuring
  the unused pods prop (TS6133).
- gpuDetail/DriverInfoPanel.tsx, gpuDetail/ProcessTable.tsx: fix relative
  import paths to shared/TechnicalAcronym, hooks/useMCP, and
  ui/StatusBadge broken by the gpuDetail/ subfolder extraction (TS2307),
  and drop the now-unused t/useTranslation in DriverInfoPanel (TS6133).
- dashboardState.actions.ts: drop duplicate unused emitCardDragged import
  and rename the unused loadDashboardData storageKey param to
  _storageKey (TS6133).
- GPUReservations.tsx: guard against handleDeleteReservation() resolving
  to undefined on its early-exit path (TS18048).
- useGPUReservationsData.ts: coalesce deleteConfirmReservation to null so
  it matches GPUReservation | null (TS2322).
- ReservationFormModal.tsx: remove unused duplicate
  DEFAULT_RESERVATION_DURATION_HOURS constant (TS6133).
- reservationForm/handleReservationSave.ts: import TFunction from
  i18next instead of react-i18next, which does not export it (TS2305).
- playwright.yml: raise the frontend Build step's Node heap limit
  (NODE_OPTIONS=--max-old-space-size=6144) as a safety margin for this
  large split PR.
