+++ web/src/components/gpu/reservationForm/handleReservationSave.ts
@@ -1,4 +1,4 @@
-import { TFunction } from 'react-i18next'
+import type { TFunction } from 'i18next'
 import {
   createOrUpdateResourceQuota,
   deleteResourceQuota,
